requirements.txt
mysql-connector-python==8.0.21
djangorestframework==3.11.1
Django==3.1.1
gunicorn==20.0.4
django-heroku==0.3.1
